(#2) Cast shade ID to integer
This attempts to cast the shade ID (conveyed via the ST deviceID) into
an integer so that eventually the JSON builder will emit a correctly
encoded value to pass to the PV hub.

Perhaps there is a better place upstream to validate this, but I'm not
entirely sure right now. Doing this here to try and get things working
for gen 2.